Render instance ID as unsigned in the logger
Reviewed-on: http://review.couchbase.org/95124
Tested-by: Build Bot <build@couchbase.com>

+++ src/couchbase/log_formatter.c
@@ -66,7 +66,7 @@ void pcbc_log_formatter(char *buf, int buf_size, const char *severity, const cha
         }
     }
     if (is_lcb) {
-        snprintf(buf, buf_size, "[cb,%s] (%s L:%d I:%d) %s", severity, subsystem, srcline, instance_id, msg);
+        snprintf(buf, buf_size, "[cb,%s] (%s L:%d I:%u) %s", severity, subsystem, srcline, (unsigned int)instance_id, msg);
     } else if (instance_ptr) {
         snprintf(buf, buf_size, "[cb,%s] (%s L:%d) %s. I=%p", severity, subsystem, srcline, msg, instance_ptr);
     } else {
